+++ framework/src/lib.rs
@@ -5,7 +5,7 @@ pub mod upgrade_fixtures;
 
 //////// 0L ///////
 /// Returns the release bundle for the current code.
-pub fn head_release_bundle() -> diem_framework::ReleaseBundle {
+pub fn testing_local_release_bundle() -> diem_framework::ReleaseBundle {
     release::ReleaseTarget::Head
         .load_bundle()
         .expect("could not find release bundle, head.mrb in framework/releases/head.mrb")
