+++ tools/genesis/tests/json_to_genesis_audit_all.rs
@@ -4,7 +4,7 @@ use diem_state_view::account_with_state_view::AsAccountWithStateView;
 use diem_storage_interface::state_view::LatestDbStateCheckpointView;
 use diem_types::{account_view::AccountView, chain_id::NamedChain};
 use libra_backwards_compatibility::legacy_recovery_v6::AccountRole;
-use libra_framework::head_release_bundle;
+use libra_framework::testing_local_release_bundle;
 use libra_genesis_tools::{
     compare,
     genesis::make_recovery_genesis_from_vec_legacy_recovery,
@@ -34,7 +34,7 @@ fn test_correct_supply_arithmetic_all() {
     let gen_tx = make_recovery_genesis_from_vec_legacy_recovery(
         &mut user_accounts,
         &genesis_vals,
-        &head_release_bundle(),
+        &testing_local_release_bundle(),
         ChainId::mainnet(),
         &libra_genesis_default(NamedChain::MAINNET),
     )
@@ -105,7 +105,7 @@ fn test_drop_all() {
     let gen_tx = make_recovery_genesis_from_vec_legacy_recovery(
         &mut user_accounts,
         &genesis_vals,
-        &head_release_bundle(),
+        &testing_local_release_bundle(),
         ChainId::mainnet(),
         &libra_genesis_default(NamedChain::MAINNET),
     )
